完善玩家征收链路（open/collect）并补齐设施收益映射与协议接线
本次提交围绕“城内征收”功能打通 player 侧完整链路，覆盖协议扩展、dispatcher 分发、handler 业务处理、设施收益计算、实体字段类型修正以及相关生成代码同步，目标是让前端 `interior.openCollect` / `interior.collect` 请求能够被稳定处理并返回可直接消费的数据。

1. player 协议扩展与生成代码同步
- 在 `player.proto` 中新增：
  - `CollectionRequest`（请求）
  - `CollectionResponse`（响应，包含 `gold/limit/curTimes/nextTime`）
- 在 `PlayerRequest`/`PlayerResponse` oneof 中新增 `collectionRequest` / `collectionResponse` 分支。
- 重新生成 `internal/shared/gen/player/player.pb.go`，同步 oneof wrapper、getter、消息索引等代码。

2. dispatcher 接线补全
- 在 `internal/player/actors/dispatcher.go` 中：
  - 注册 `HandleCollectionRequest`
  - 在 `unwrapPlayerRequestBody` 增加 `CollectionRequest` 解包分支
- 确保 `collection` 请求可从 actor 分发层正确路由到 handler。

3. OpenCollection 业务完善（面板打开）
- 在 `HandleOpenCollectionRequest` 中完成次数/时间计算逻辑：
  - 读取配置 `CollectInterval` 与 `CollectTimesLimit`
  - 计算当前可领取状态与下次领取时间
  - 当日未领取时返回 `nextTime=0`
  - 达上限时返回次日零点时间
- 增加玩家属性空指针保护，避免属性未初始化时 panic。

4. Collection 业务实现（执行领取）
- 新增 `HandleCollectionRequest`：
  - 校验每日次数上限
  - 校验冷却时间（CD）
  - 计算设施收益并返回本次金币产出
  - 更新征收状态（`lastCollectTime`、`collectTimes`）
- 返回统一 `PlayerResponse_CollectionResponse`，保持与现有 handler 响应风格一致。

5. 设施收益计算能力补齐
- 在 `player_service.go` 新增 `ComputeFacilityYield`，按玩家设施等级汇总资源/金币收益。
- 在 `facility` 配置模块中增强：
  - 设施明细索引字段公开为 `Facilities`
  - 新增 `Facility.LevelMap`（等级 -> 配置）索引
  - 加载时构建 `LevelMap` 并做重复等级校验
  - `GetFacilityYield` 增加空值与等级缺失防御，避免越界/空 map 触发异常。

6. 设施等级字段类型修正
- 将设施等级字段 `privateLevel` 从 `int8` 调整为 `int`，并同步实体生成代码与持久化模型：
  - `internal/player/entity/domain/facility.go`
  - `internal/player/entity/facility_entity_gen.go`
  - `internal/player/infra/persistence/model/facility_doc_gen.go`
- 目的是统一与配置等级索引（`int`）的类型语义，减少转换与比较风险。

7. 清理冲突的 account 生成物
- 删除 `internal/shared/gen/account/role.pb.go`，消除与 `account.pb.go` 中重复类型定义带来的编译冲突。

整体效果
- player 侧征收能力已形成“打开面板（open）+ 执行征收（collect）”的完整协议与处理链路。
- 设施收益计算具备可用的等级索引与容错能力，为征收收益返回提供稳定基础。
- actor 分发、proto 与生成代码保持一致，降低后续联调时的类型错配风险。

+++ internal/player/infra/persistence/model/facility_doc_gen.go
@@ -8,7 +8,7 @@ import (
 type FacilityDoc struct {
 	Id           int    `bson:"id"`
 	Name         string `bson:"name"`
-	PrivateLevel int8   `bson:"private_level"`
+	PrivateLevel int    `bson:"private_level"`
 	FType        int8   `bson:"f_type"`
 	UpTime       int64  `bson:"up_time"`
 }
